LayoutTests/platform/gtk/css-dark-mode/prefers-color-scheme-expected.txt


PASS Element colors are correct in light color scheme 
PASS Element colors are correct in light color scheme with boolean context query 
PASS Dark color scheme enabled 
FAIL Element colors are correct in dark color scheme assert_equals: expected "rgb(0, 0, 0)" but got "rgb(255, 255, 255)"
FAIL Element colors are correct in dark color scheme with boolean context query assert_equals: expected "rgb(0, 0, 0)" but got "rgb(255, 255, 255)"
